frontend/requirements.txt
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
streamlit-option-menu
streamlit-autorefresh
plotly
pandas
